feat: Added devel-required-configs to make poorly-featured libcs easier to work with
A load of other things were also found while doing this:
- Had to up the file size limit to 175KB because of the unused gnulib-based code in vasnprintf
- Added a few more comments to CMakeLists.txt
- Fixed a few check_libc_symbol calls in CMakeLists.txt to use the
  standard-specific calls when appropriate
- To make dietlibc work better a few new portable-symbols were added,
  those being dprintf, vdprintf, full_write, safe_write, vasnprintf
  and SYS_BUFSIZE_MAX
- A large amount of typos and other small errors in many comments were
  fixed
- Some spacing issues were resolved
- A few feature tests were fixed to cover more things (in particular
  YALIBCT_DISABLE_RARE_LOCALE_TESTS)
- picolibc_scanf.c was updated to contain an upstream change

+++ tests/printf/picolibc_scanf.c
@@ -50,7 +50,7 @@
 # define _WANT_IO_C99_FORMATS
 # define _WANT_IO_LONG_LONG
 # define _WANT_IO_POS_ARGS
-# define printf_float(x) (x)
+# define printf_float(x) ((double) (x))
 
 #ifdef YALIBCT_DISABLE_PRINTF_NUMBERED_ARGUMENTS_TESTS
 #define NO_POS_ARGS
